+++ protostar/commands/migrate/migrate_command.py
@@ -8,10 +8,10 @@ from starknet_py.net.signer import BaseSigner
 from protostar.cli import Command
 from protostar.cli.network_command_util import NetworkCommandUtil
 from protostar.cli.signable_command_util import SignableCommandUtil
-from protostar.commands.test.test_environment_exceptions import CheatcodeException
 from protostar.migrator import Migrator, MigratorExecutionEnvironment
 from protostar.protostar_exception import ProtostarException
 from protostar.starknet_gateway.gateway_facade_factory import GatewayFacadeFactory
+from protostar.test_runner.test_environment_exceptions import CheatcodeException
 from protostar.utils.input_requester import InputRequester
 from protostar.utils.log_color_provider import LogColorProvider
