+++ protostar/commands/migrate/migrate_command_test.py
@@ -8,10 +8,10 @@ from starknet_py.net.gateway_client import GatewayClient
 from starknet_py.net.models import StarknetChainId
 
 from protostar.commands.migrate.migrate_command import MigrateCommand
-from protostar.commands.test.test_environment_exceptions import CheatcodeException
 from protostar.migrator import Migrator
 from protostar.protostar_exception import ProtostarException
 from protostar.starknet_gateway import GatewayFacadeFactory
+from protostar.test_runner.test_environment_exceptions import CheatcodeException
 from protostar.utils.input_requester import InputRequester
