+++ protostar/migrator/cheatcodes/migrator_call_cheatcode.py
@@ -3,13 +3,13 @@ from typing import Any, Optional
 
 from typing_extensions import Protocol
 
-from protostar.commands.test.test_environment_exceptions import CheatcodeException
 from protostar.starknet.cheatcode import Cheatcode
 from protostar.starknet_gateway import (
     ContractNotFoundException,
     GatewayFacade,
     UnknownFunctionException,
 )
+from protostar.test_runner.test_environment_exceptions import CheatcodeException
 from protostar.utils.data_transformer import CairoOrPythonData
