+++ protostar/migrator/cheatcodes/migrator_deploy_contract_cheatcode.py
@@ -5,14 +5,14 @@ from typing import Any, Optional
 
 from typing_extensions import Protocol
 
-from protostar.commands.test.test_environment_exceptions import (
-    KeywordOnlyArgumentCheatcodeException,
-)
 from protostar.compiler import ProjectCompiler
 from protostar.compiler.compiled_contract_writer import CompiledContractWriter
 from protostar.migrator.migrator_datetime_state import MigratorDateTimeState
 from protostar.starknet.cheatcode import Cheatcode
 from protostar.starknet_gateway.gateway_facade import GatewayFacade
+from protostar.test_runner.test_environment_exceptions import (
+    KeywordOnlyArgumentCheatcodeException,
+)
 from protostar.utils.data_transformer import CairoOrPythonData
 
 from .network_config import CheatcodeNetworkConfig, ValidatedCheatcodeNetworkConfig
