+++ protostar/test_runner/cheatcodes/load_cheatcode.py
@@ -2,9 +2,9 @@ from typing import Any, Callable, List, Optional
 
 from starkware.starknet.business_logic.state.state import ContractStorageState
 
-from protostar.commands.test.test_environment_exceptions import CheatcodeException
 from protostar.starknet.cheatcode import Cheatcode
 from protostar.starknet.storage_var import calc_address
+from protostar.test_runner.test_environment_exceptions import CheatcodeException
 
 
 class LoadCheatcode(Cheatcode):
