+++ protostar/test_runner/cheatcodes/mock_call_cheatcode.py
@@ -1,12 +1,11 @@
 from collections.abc import Mapping
 from typing import Any, Callable, List, Optional
 
-from starkware.starknet.public.abi import AbiType
-from starkware.starknet.public.abi import get_selector_from_name
+from starkware.starknet.public.abi import AbiType, get_selector_from_name
 
-from protostar.commands.test.test_environment_exceptions import CheatcodeException
 from protostar.starknet.cheatcode import Cheatcode
 from protostar.starknet.types import AddressType
+from protostar.test_runner.test_environment_exceptions import CheatcodeException
 from protostar.utils.data_transformer import (
     CairoOrPythonData,
     PythonData,
