+++ protostar/test_runner/environments/fuzz_test_execution_environment.py
@@ -10,10 +10,6 @@ from hypothesis.reporting import with_reporter
 from hypothesis.strategies import SearchStrategy
 from starkware.starknet.business_logic.execution.objects import CallInfo
 
-from protostar.commands.test.test_environment_exceptions import (
-    CheatcodeException,
-    ReportedException,
-)
 from protostar.starknet.cheatcode import Cheatcode
 from protostar.test_runner.cheatcodes import AssumeCheatcode, RejectCheatcode
 from protostar.test_runner.fuzzing.exceptions import HypothesisRejectException
@@ -31,6 +27,10 @@ from protostar.test_runner.starkware.execution_resources_summary import (
     ExecutionResourcesSummary,
 )
 from protostar.test_runner.starkware.test_execution_state import TestExecutionState
+from protostar.test_runner.test_environment_exceptions import (
+    CheatcodeException,
+    ReportedException,
+)
 from protostar.utils.abi import get_function_parameters
 
 from .test_execution_environment import (
