+++ protostar/test_runner/test_case_runners/fuzz_test_case_runner.py
@@ -1,15 +1,15 @@
 from typing import Optional
 
-from protostar.commands.test.test_environment_exceptions import (
-    BreakingReportedException,
-    ReportedException,
-)
 from protostar.test_runner.environments.fuzz_test_execution_environment import (
     FuzzTestExecutionEnvironment,
     FuzzTestExecutionResult,
 )
 from protostar.test_runner.fuzzing.fuzz_input_exception_metadata import (
     FuzzInputExceptionMetadata,
+)
+from protostar.test_runner.test_environment_exceptions import (
+    BreakingReportedException,
+    ReportedException,
 )
 from protostar.test_runner.test_results import (
     BrokenFuzzTestCaseResult,
