+++ protostar/test_runner/test_context.py
@@ -1,8 +1,9 @@
 from types import SimpleNamespace
 from typing import Any
 
-from protostar.commands.test.test_environment_exceptions import SimpleReportedException
 from protostar.starknet.hint_local import HintLocal
+
+from .test_environment_exceptions import SimpleReportedException
 
 
 class TestContext(SimpleNamespace):
