+++ protostar/test_runner/test_environment_exceptions.py
@@ -6,8 +6,9 @@ from starkware.starknet.business_logic.execution.objects import Event
 from typing_extensions import Literal
 
 from protostar.starknet.hint_local import HintLocal
-from protostar.test_runner.expected_event import ExpectedEvent
 from protostar.utils.log_color_provider import SupportedColorName, log_color_provider
+
+from .expected_event import ExpectedEvent
 
 # NOTE: When adding new exception type, do not forget to include it in ``test_pickle`` test.
